requirements.txt
torch>=2.1
torchvision>=0.16
# Install the Ascend-matched torch_npu wheel separately in the active NPU environment.
pandas>=2.0
numpy>=1.24
pvlib>=0.10
Pillow>=10.0
pyyaml>=6.0
tqdm>=4.66
